fix: return integer

+++ src/pytemplate/service/calculator.py
@@ -15,7 +15,7 @@ class Calculator:
         return operands.first_operand * operands.second_operand
 
     @staticmethod
-    def divide(operands: Operands) -> float:
+    def divide(operands: Operands) -> int:
         if operands.second_operand == 0:
             raise ValueError("Cannot divide by zero")
         return operands.first_operand / operands.second_operand
